Yetkilendirme işlemi yapıldı.

+++ MealOrdering/Server/Services/Services/OrderService.cs
@@ -16,11 +16,13 @@ namespace MealOrdering.Server.Services.Services
     {
         private readonly MealOrderingDbContext context;
         private readonly IMapper mapper;
+        private readonly IValidationService validationService;
 
-        public OrderService(MealOrderingDbContext context, IMapper mapper)
+        public OrderService(MealOrderingDbContext context, IMapper mapper, IValidationService validationService)
         {
             this.context = context;
             this.mapper = mapper;
+            this.validationService = validationService;
         }
 
         public async Task<OrderDTO> CreateOrder(OrderDTO Order)
@@ -58,11 +60,16 @@ namespace MealOrdering.Server.Services.Services
             var detailCount = await context.OrderItems.Where(i => i.OrderId == OrderId).CountAsync();
 
             if (detailCount > 0)
-                throw new Exception($"Bu siparişe ait {detailCount} adet alt sipariş var.");
+                throw new Exception($"There are {detailCount} sub items for the order you are trying to delete");
 
             var order = await context.Orders.FirstOrDefaultAsync(i => i.Id == OrderId);
             if (order == null)
-                throw new Exception("Sipariş bulunamadı");
+                throw new Exception("Order not found");
+
+            if (!validationService.HasPermission(order.CreatedUserId))
+                throw new Exception("You cannot change the order unless you created");
+
+
 
             context.Orders.Remove(order);
 
@@ -139,6 +146,9 @@ namespace MealOrdering.Server.Services.Services
             var dbOrder = await context.Orders.FirstOrDefaultAsync(i => i.Id == Order.Id);
             if (dbOrder == null)
                 throw new Exception("Sipariş Bulunamadı");
+
+            if (!validationService.HasPermission(dbOrder.CreatedUserId))
+                throw new Exception("You cannot change the order unless you created");
 
             mapper.Map(Order, dbOrder);
             await context.SaveChangesAsync();
